salesperson-portal/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.2
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
redis==5.0.1
celery==5.3.6
opencv-python-headless==4.8.1.78
ultralytics==8.0.222
cryptography==41.0.7
minio==7.2.0
python-dotenv==1.0.0
alembic==1.13.0
